requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
configobj==5.0.8
decorator==4.4.2
frozenlist==1.4.0
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
magic-filter==1.0.11
moviepy==1.0.3
multidict==6.0.4
numpy==1.24.4
openai==0.28.1
opencv-python==4.8.1.78
Pillow==10.0.1
proglog==0.1.10
pydantic==2.3.0
pydantic_core==2.6.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==5.0.1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
